should be able to edit a job; radiobuttons and checkboxes need work
