+++ internify/backend/routes/api/jobs.js
@@ -58,24 +58,17 @@ router.get("/:id", function (req, res, next) {
 //@access   Public
 router.put("/:id", function (req, res, next) {
   const jobId = req.params.id;
-  JobPostingData.findOne({ jobId: jobId })
-    .then((job) => {
-      job.dateUpdated = req.body.dateUpdated
-        ? req.body.dateUpdated
-        : job.dateUpdated;
-      job.score = req.body.score ? req.body.score : job.score;
-      job.header = req.body.header ? req.body.header : job.header;
-      job.requirements = req.body.requirements
-        ? req.body.requirements
-        : job.requirements;
-      job.details = req.body.details ? req.body.details : job.details;
-      job.contact = req.body.contact ? req.body.contact : job.contact;
-
-      job.save().then(() => {
-        JobPostingData.find().then((jobs) => res.json(jobs));
-      });
-    })
-    .catch((err) => res.status(404).json({ success: false }));
+  var args = {
+    score: req.body.score,
+    header: req.body.header,
+    requirements: req.body.requirements,
+    details: req.body.details,
+    contact: req.body.contact,
+  };
+  JobPostingData.findByIdAndUpdate(
+    jobId, { $set: { ...args, }, }
+  )
+  .catch((err) => res.status(404).json({ success: false }));
 });
 
 //@route    POST api/jobs
