+++ internify/src/components/pages/Edit.js
@@ -23,6 +23,7 @@ import { processMatches } from "../../store/actions/matchesActions";
 import axios from "axios";
 import firebase from "firebase/app";
 import "firebase/auth";
+import { createJobObject, checkIfEmpty } from "../../effects/filter.effects";
 
 const mockTechStackData = {
   languages: ["Java", "JavaScript", "C++", "C"],
@@ -79,104 +80,9 @@ function Edit(props) {
   const page1Object = useSelector((state) => state.matches.page1Object);
   const page2Object = useSelector((state) => state.matches.page2Object);
 
-  function parseConcepts(concepts) {
-    let parsedConcepts = [];
-    for (let i = 0; i < concepts.length; i++) {
-      if (concepts[i] === "Object Oriented Programming") {
-        parsedConcepts.push("Object-Oriented Programming");
-      } else {
-        parsedConcepts.push(concepts[i]);
-      }
-    }
-    return parsedConcepts;
-  }
-
-  function parseExperience(experience) {
-    if (experience === "none") {
-      return 0;
-    } else {
-      const expArray = experience.split(" ");
-      return parseInt(expArray[1]);
-    }
-  }
-
-  function parseCitizenshipReqs(candidates) {
-    let reqs = [];
-    // anyone case
-    if (candidates === "Anyone") {
-      reqs = ["Anyone"]; // always return true in checker for this!
-      return reqs;
-    } else {
-      // logic-handling for other cases
-      const expArray = candidates.split(" ");
-      if (expArray.includes("Citizens") && expArray.includes("PR")) {
-        reqs = ["Citizen", "Permanent Residency"];
-        return reqs;
-      }
-      if (expArray.includes("Citizens")) {
-        reqs = ["Citizen"];
-        return reqs;
-      }
-      reqs = ["Citizen", "Permanent Residency", "International"];
-      return reqs;
-    }
-  }
-
-  function parseCoopReqs(coOp) {
-    if (coOp === "Yes") {
-      return true;
-    } else {
-      return false;
-    }
-  }
-
-  function parseLanguages(languages) {
-    let editedLanguages = [];
-    for (let i = 0; i < languages.length; i++) {
-      if (languages[i] === "C++") {
-        editedLanguages.push("Cpp");
-        editedLanguages.push("C++");
-      } else if (languages[i] === "Cpp") {
-        editedLanguages.push("C++");
-        editedLanguages.push("Cpp");
-      } else {
-        editedLanguages.push(languages[i]);
-      }
-    }
-    return editedLanguages;
-  }
-
-  function editJobObject(jobPosting) {
-    // parsing functions
-    const experience = parseExperience(jobPosting.requirements.experience);
-    const candidates = parseCitizenshipReqs(jobPosting.details.candidates);
-    const coOp = parseCoopReqs(jobPosting.details.coOp);
-    const concepts = parseConcepts(jobPosting.requirements.concepts);
-    const languages = parseLanguages(jobPosting.requirements.languages);
-
-    return {
-      jobId: jobPosting.jobId,
-      experience: experience,
-      gpa: jobPosting.requirements.gpa,
-      gpaValue: jobPosting.requirements.gpaValue,
-      languages: languages,
-      frameworks: jobPosting.requirements.frameworks,
-      tools: jobPosting.requirements.tools,
-      concepts: concepts,
-      candidates: candidates,
-      academicReq: jobPosting.details.academicReq,
-      coOp: coOp,
-    };
-  }
-  function checkIfEmpty(obj) {
-    const sub = jobData[obj];
-    for (var key in sub) {
-      const currVal = sub[key];
-      if (currVal === "" || currVal.length === 0) {
-        return true;
-      }
-    }
-    return false;
+  function editJob(data, jobId, props) {
+    props.actions.editJobsData(data);
+    window.scrollTo(0, 0);
   }
 
   useEffect(() => {
@@ -188,11 +94,11 @@ function Edit(props) {
     window.scrollTo(0, 0);
     const curr = currStep[currentStep];
 
-    if (!checkIfEmpty(curr)) {
+    if (!checkIfEmpty(curr, jobData)) {
       setError(false);
 
       props.actions.editJobsData(jobData);
-      const posting = editJobObject(jobData);
+      const posting = createJobObject(jobData);
 
       if (currentStep === 1) {
         props.actions.processMatches({
@@ -255,7 +161,7 @@ function Edit(props) {
             currentStep={currentStep}
             handleChange={setJobData}
             jobData={jobData}
-            pathway={"edit"}
+            title={"1. Edit Job Header"}
           />
           <TechRequirements
             currentStep={currentStep}
@@ -274,8 +180,8 @@ function Edit(props) {
             handleChange={setJobData}
             jobData={jobData}
             data={mockJobDetailData}
-          />
-          <Review currentStep={currentStep} jobData={jobData} user={user} />
+          /> 
+          <Review currentStep={currentStep} jobData={jobData} user={user} onSubmit={editJob} />
           {currentStep < 5 ? (
             <Container maxWidth="md">
               <ButtonFilled onClick={() => updateStore()}>
